tests/wpt/mozilla/tests/input-events/accesskey-on-html-element.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>The accesskey attribute should work properly on focusable divs</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/accesskey.js"></script>
<div style="height: 10000px"></div>
<div id="target" accesskey="P" tabindex="0">Access Key</div>

<script>
promise_test(async function() {
    await pressAccessKey('P');
    let boundingRect = target.getBoundingClientRect();
    assert_true(boundingRect.x > 0);
    assert_true(boundingRect.x < document.scrollingElement.clientWidth);
    assert_true(boundingRect.y > 0);
    assert_true(boundingRect.y < document.scrollingElement.clientHeight);
});
</script>
